+++ backend/src/event/EVENT_CONTRIBUTION_CREATE.ts
@@ -1,0 +1,12 @@
+import Decimal from 'decimal.js-light'
+import { User as DbUser } from '@entity/User'
+import { Contribution as DbContribution } from '@entity/Contribution'
+import { Event as DbEvent } from '@entity/Event'
+import { Event, EventType } from './Event'
+
+export const EVENT_CONTRIBUTION_CREATE = async (
+  user: DbUser,
+  contribution: DbContribution,
+  amount: Decimal,
+): Promise<DbEvent> =>
+  Event(EventType.CONTRIBUTION_CREATE, user, user, null, null, contribution, null, amount).save()
