+++ backend/src/graphql/resolver/ContributionResolver.test.ts
@@ -45,7 +45,7 @@ import { Event as DbEvent } from '@entity/Event'
 import { Contribution } from '@entity/Contribution'
 import { Transaction as DbTransaction } from '@entity/Transaction'
 import { User } from '@entity/User'
-import { EventType } from '@/event/EventType'
+import { EventType } from '@/event/Event'
 import { logger, i18n as localization } from '@test/testSetup'
 import { UserInputError } from 'apollo-server-express'
 import { raeuberHotzenplotz } from '@/seeds/users/raeuber-hotzenplotz'
