+++ backend/src/graphql/resolver/UserResolver.test.ts
@@ -34,7 +34,7 @@ import { contributionLinkFactory } from '@/seeds/factory/contributionLink'
 import { transactionLinkFactory } from '@/seeds/factory/transactionLink'
 import { ContributionLink } from '@model/ContributionLink'
 import { TransactionLink } from '@entity/TransactionLink'
-import { EventType } from '@/event/EventType'
+import { EventType } from '@/event/Event'
 import { Event as DbEvent } from '@entity/Event'
 import { validate as validateUUID, version as versionUUID } from 'uuid'
 import { peterLustig } from '@/seeds/users/peter-lustig'
